+++ packages/themes/CHANGELOG.md
@@ -1,5 +1,12 @@
 # @getdashfy/themes
 
+## 0.1.3
+
+### Patch Changes
+
+- Updated dependencies []:
+  - @getdashfy/types@0.3.0
+
 ## 0.1.2
 
 ### Patch Changes