+++ packages/ui/CHANGELOG.md
@@ -1,5 +1,12 @@
 # @getdashfy/ui
 
+## 0.3.1
+
+### Patch Changes
+
+- Updated dependencies []:
+  - @getdashfy/themes@0.1.3
+
 ## 0.3.0
 
 ### Minor Changes